+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-database/src/main/java/com/microsoft/azure/toolkit/intellij/database/mysql/connection/MySqlDatabaseResourceDefinition.java
@@ -21,6 +21,7 @@ import javax.annotation.Nonnull;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Stream;
 
 public class MySqlDatabaseResourceDefinition extends SqlDatabaseResourceDefinition<MySqlDatabase> {
     public static final MySqlDatabaseResourceDefinition INSTANCE = new MySqlDatabaseResourceDefinition();
@@ -32,6 +33,20 @@ public class MySqlDatabaseResourceDefinition extends SqlDatabaseResourceDefiniti
     @Override
     public MySqlDatabase getResource(String dataId) {
         return Azure.az(AzureMySql.class).getById(dataId);
+    }
+
+    @Override
+    public List<Resource<MySqlDatabase>> getResources(Project project) {
+        return Azure.az(AzureMySql.class).list().stream()
+            .flatMap(m -> m.servers().list().stream())
+            .flatMap(s -> {
+                try {
+                    return s.databases().list().stream();
+                } catch (final Throwable e) {
+                    return Stream.empty();
+                }
+            })
+            .map(this::define).toList();
     }
 
     @Override