+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-service-explorer/src/main/java/com/microsoft/azure/toolkit/intellij/explorer/action/AzureExplorerActionsContributor.java
@@ -1,7 +1,6 @@
 package com.microsoft.azure.toolkit.intellij.explorer.action;
 
 import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.openapi.wm.ToolWindowManager;
 import com.microsoft.azure.toolkit.ide.common.IActionsContributor;
 import com.microsoft.azure.toolkit.ide.common.action.ResourceCommonActionsContributor;
@@ -9,22 +8,17 @@ import com.microsoft.azure.toolkit.intellij.explorer.AzureExplorer;
 import com.microsoft.azure.toolkit.lib.common.action.AzureActionManager;
 import com.microsoft.azure.toolkit.lib.common.task.AzureTaskManager;
 
-import java.util.Objects;
+import java.util.Optional;
 import java.util.function.BiConsumer;
 
 public class AzureExplorerActionsContributor implements IActionsContributor {
     @Override
     public void registerHandlers(AzureActionManager am) {
-        final BiConsumer<Object, AnActionEvent> openAzureExplorer = (ignore, e) -> openAzureExplorer(e);
+        final BiConsumer<Object, AnActionEvent> openAzureExplorer = (ignore, e) -> Optional.ofNullable(e).map(AnActionEvent::getProject)
+            .map(ToolWindowManager::getInstance)
+            .map(m -> m.getToolWindow(AzureExplorer.TOOLWINDOW_ID))
+            .ifPresent(w -> AzureTaskManager.getInstance().runLater(() -> w.show(null)));
         am.registerHandler(ResourceCommonActionsContributor.OPEN_AZURE_EXPLORER, (i, e) -> true, openAzureExplorer);
-    }
-
-    private static void openAzureExplorer(AnActionEvent e) {
-        final ToolWindow toolWindow = ToolWindowManager.getInstance(Objects.requireNonNull(e.getProject())).
-            getToolWindow(AzureExplorer.TOOLWINDOW_ID);
-        if (Objects.nonNull(toolWindow) && !toolWindow.isVisible()) {
-            AzureTaskManager.getInstance().runLater((Runnable) toolWindow::show);
-        }
     }
 
     @Override