+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib-java/src/main/java/com/microsoft/azure/toolkit/intellij/connector/spring/SpringSupported.java
@@ -11,18 +11,45 @@ import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 public interface SpringSupported<T> extends ResourceDefinition<T> {
-    static List<Pair<String, String>> getProperties(Connection<?, ?> c) {
+    static List<Pair<String, String>> getProperties(Connection<?, ?> c, final String propKey) {
         final ResourceDefinition<?> rd = c.getResource().getDefinition();
         if (rd instanceof SpringSupported) {
-            return ((SpringSupported<?>) rd).getSpringProperties().stream()
-                    .map(p -> Pair.of(p.getKey(), p.getValue().replaceAll(Connection.ENV_PREFIX, c.getEnvPrefix())))
-                    .collect(Collectors.toList());
+            return ((SpringSupported<?>) rd).getSpringProperties(propKey).stream()
+                .map(p -> Pair.of(p.getKey(), p.getValue().replaceAll(Connection.ENV_PREFIX, c.getEnvPrefix())))
+                .collect(Collectors.toList());
         }
         return Collections.emptyList();
     }
 
-    List<Pair<String, String>> getSpringProperties();
+    static Map<String, String> getPropertyMethods(Connection<?, ?> c) {
+        final ResourceDefinition<?> rd = c.getResource().getDefinition();
+        if (rd instanceof SpringSupported) {
+            return ((SpringSupported<?>) rd).getSpringPropertyFields();
+        }
+        return Collections.emptyMap();
+    }
+
+    List<Pair<String, String>> getSpringProperties(String key);
+
+    default List<Pair<String, String>> getSpringProperties() {
+        return getSpringProperties(null);
+    }
+
+    /**
+     * get the binding fields of configuration properties.
+     */
+    default Map<String, String> getSpringPropertyFields() {
+        return Collections.emptyMap();
+    }
+
+    /**
+     * get value types of configuration properties.
+     */
+    default Map<String, String> getSpringPropertyTypes() {
+        return Collections.emptyMap();
+    }
 }